refactor(data): conform external resolver to the namespace file/export standard
Rename external.ts -> resolve-external-invocation.ts so the filename maps
deterministically to its single export (do-bar.ts -> doBar), and drop the
non-standard flat re-export from schema/index.ts — reach it via the Schema
namespace (Schema.resolveExternalInvocation), per data-ai global/namespace.md.
