+++ packages/data/src/schema/index.ts
@@ -17,7 +17,6 @@ export * from "./boolean/index.js";
 export { F32, I32, U32, F64 } from "../math/index.js";
 export * from "./time/index.js";
 export { toVertexBufferLayout, toVertexBufferLayoutForType } from "./to-vertex-buffer-layout.js";
-export { resolveExternalInvocation } from "./external.js";
 export type { GPUVertexBufferLayout, GPUVertexAttributeDescriptor, GPUVertexFormat } from "./to-vertex-buffer-layout.js";
 export * from "./fractional-index/fractional-index.js";
 export * from "./guid/index.js";
