+++ packages/data/src/schema/public.ts
@@ -4,7 +4,7 @@ export * from "./schema.js";
 export * from "./to-vertex-buffer-layout.js";
 export * from "./nullable.js";
 export * from "./to-type.js";
-export * from "./external.js";
+export * from "./resolve-external-invocation.js";
 export * from "./from-object-properties.js";
 export * from "./from-archetype.js";
 export * from "./from-struct-properties.js";
