+++ packages/data/src/schema/resolve-external-invocation.ts
